requirements.txt
torch
torchvision
transformers
pillow
opencv-python
matplotlib
streamlit
numpy
timm
ftfy
streamlit-image-comparison